доаботал логирование, рефакторинг request

+++ utils/utils_bot.py
@@ -2,7 +2,11 @@ import asyncio
 from aiogram import Bot
 from typing import List
 from aiogram.exceptions import TelegramAPIError
-from config import config
+from config import Config
+import logging
+
+
+logger = logging.getLogger("bot")
 
 
 async def message_deleter(
@@ -20,16 +24,19 @@ async def message_deleter(
     :param delay: Задержка в секундах. Если не указано — берётся из config.TEMP_MESSAGE_LIFETIME_SEC
     """
 
-    actual_delay = delay if delay is not None else config.TEMP_MESSAGE_LIFETIME_SEC
+    if not message_ids:
+        return
+
+    actual_delay = delay if delay is not None else Config.TEMP_MESSAGE_LIFETIME_SEC
     await asyncio.sleep(actual_delay)
 
     for msg_id in message_ids:
         try:
             await bot.delete_message(chat_id=chat_id, message_id=msg_id)
         except TelegramAPIError:
-            # Игнорируем ошибки Telegram API:
-            # - сообщение уже удалено
-            # - прошло более 48 часов (в ЛС)
-            # - нет прав на удаление
-            # - пользователь заблокировал бота и т.п.
             pass
+        except Exception as e:
+            logger.error(
+                f"Неожиданная ошибка при удалении {msg_id} в чате {chat_id}: {e}",
+                exc_info=True
+            )
